rename loader to module
