+++ pyk4a/pyk4a.py
@@ -5,7 +5,7 @@ from .calibration import Calibration
 from .capture import PyK4ACapture
 from .config import ColorControlCommand, ColorControlMode, Config
 from .errors import K4AException, _verify_error
-from .loader import k4a_module
+from .module import k4a_module
 
 
 if sys.version_info < (3, 8):
